Move tests for operation details into single file
The tests that were in each file for the different operations were
actually testing the details parsing and not the structs explicitly.
Since the tests mostly involve the entire operation module, I'm moved
them into their own file `test.rs` that is only built in test mode.
