+++ resources/src/operation/allow_trust.rs
@@ -49,32 +49,3 @@ impl AllowTrust {
         self.authorize
     }
 }
-
-#[cfg(test)]
-mod allow_trust_tests {
-    use serde_json;
-    use operation::{Operation, OperationDetail};
-
-    fn change_trust_json() -> &'static str {
-        include_str!("../../fixtures/operations/allow_trust.json")
-    }
-
-    #[test]
-    fn it_parses_change_trust_from_json() {
-        let operation: Operation = serde_json::from_str(&change_trust_json()).unwrap();
-        assert!(operation.is_allow_trust());
-        assert_eq!(operation.type_i(), 7);
-        if let &OperationDetail::AllowTrust(ref account_details) = operation.detail() {
-            assert_eq!(
-                account_details.trustee(),
-                "GC23QF2HUE52AMXUFUH3AYJAXXGXXV2VHXYYR6EYXETPKDXZSAW67XO4"
-            );
-            assert_eq!(
-                account_details.trustor(),
-                "GBXGQJWVLWOYHFLVTKWV5FGHA3LNYY2JQKM7OAJAUEQFU6LPCSEFVXON"
-            );
-            assert_eq!(account_details.asset().code(), "USD");
-            assert_eq!(account_details.authorize(), true);
-        }
-    }
-}
