+++ resources/src/operation/manage_data.rs
@@ -1,5 +1,4 @@
 /// Set, modify or delete a Data Entry (name/value pair) for an account.
-
 #[derive(Debug, Deserialize)]
 pub struct ManageData {
     name: String,
@@ -23,26 +22,5 @@ impl ManageData {
     /// The new data value associated with the named key
     pub fn value(&self) -> &String {
         &self.value
-    }
-}
-
-#[cfg(test)]
-mod manage_data_tests {
-    use serde_json;
-    use operation::{Operation, OperationDetail};
-
-    fn manage_data_json() -> &'static str {
-        include_str!("../../fixtures/operations/manage_data.json")
-    }
-
-    #[test]
-    fn it_parses_manage_data_from_json() {
-        let operation: Operation = serde_json::from_str(&manage_data_json()).unwrap();
-        assert!(operation.is_manage_data());
-        assert_eq!(operation.type_i(), 10);
-        if let &OperationDetail::ManageData(ref account_details) = operation.detail() {
-            assert_eq!(account_details.name(), "lang");
-            assert_eq!(account_details.value(), "aW5kb25lc2lhbg==");
-        }
     }
 }
